feat
Added Postgres local database and created the pipeline from frontend to database.

The frontend now creates messages served up from data storted in a Postgres database. Flask now takes new post data from the frontend and sends it straight back to the database. When the server starts up, flask will retreive that database info and send it to frontend to be rendered.